extract dataclasses for electricity2018
